refactor: move ExprFilter to dedicated file
Address review comment: extract ExprFilter from engine.rs into
filter/expr_filter.rs for better separation of concerns.
