+++ crates/scouty/src/filter/mod.rs
@@ -1,3 +1,4 @@
 pub mod engine;
 pub mod expr;
 pub mod eval;
+pub mod expr_filter;
